[sparc] Add base for sparc cpu support
This adds boilerplate for sparc cpu support. This implements the sparc 2
instructions pipeline, its registers and the fetch/decode/exec
machinery. The decode.c implements a base for decoding and the whole 3
different instruction formats.

Adding a new instruction support, should only be a matter to fill the
proper static arrays in decoder.c and add code simulating the actual
instruction behavior in iu.c.