[sparc] Add cpu branch ability
This adds a new virtual register among pc and npc that can be set
during instruction execution. This can be viewed as a temporary nnpc
used to store the new npc value at the end of the instruction process.
Thus instruction handlers can change this register value to simulate a
delay jump.

This also adds helper to get pc and npc and to delay jump to a new
address.