[sparc] Do not use register address directly
Giving the general purpose register addresses to user is not a very
safe API.

This separates generic register getters and setters allowing to check
for register index boundaries. This also provides a way to secure
write access to %g0 register, that is a null only register.

This also adds a test case that checks %g0 remains null.