[sparc] Rework instruction execution logic
The iu.c file used a whole bunch of macro to simulate template functions
generation. This produced a hard to read file and hard to debug
applications.

This file has been rewritten using proper function callbacks that make
it easier to debug. The downside is that it is a little bit less
straightforward to add new instructions. The file is also a little bit
bigger, and all load/store instruction handlers code is not easily
factorizable without using macro.